etl/pyproject.toml
[tool.poetry]
name = "postgres_to_es"
version = "1.0"
description = ""
authors = ["bekishev04 <bekishev04@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
elasticsearch = "^7.17.6"
psycopg2 = "^2.9.3"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"
loguru = "0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
